requirements.txt
prophet.
pystan
yfinance 
plotly
streamlit

 